+++ mockserver-netty/src/test/java/org/mockserver/netty/integration/tls/inbound/CustomTLSProtocolsMockingIntegrationTest.java
@@ -1,0 +1,217 @@
+package org.mockserver.netty.integration.tls.inbound;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.mockserver.integration.ClientAndServer;
+import org.mockserver.mock.Expectation;
+import org.mockserver.model.HttpForward;
+import org.mockserver.testing.integration.mock.AbstractBasicMockingSameJVMIntegrationTest;
+import org.mockserver.testing.integration.mock.AbstractMockingIntegrationTestBase;
+
+import javax.net.ssl.SSLContext;
+
+import java.security.NoSuchAlgorithmException;
+import java.util.Arrays;
+
+import static junit.framework.TestCase.assertEquals;
+import static org.hamcrest.CoreMatchers.*;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasItemInArray;
+import static org.hamcrest.core.Is.is;
+import static org.junit.Assume.assumeThat;
+import static org.mockserver.configuration.ConfigurationProperties.tlsProtocols;
+import static org.mockserver.model.Header.header;
+import static org.mockserver.model.HttpForward.forward;
+import static org.mockserver.model.HttpRequest.request;
+import static org.mockserver.model.HttpResponse.response;
+import static org.mockserver.model.HttpStatusCode.OK_200;
+import static org.mockserver.stop.Stop.stopQuietly;
+
+/**
+ * @author jamesdbloom
+ */
+public class CustomTLSProtocolsMockingIntegrationTest extends AbstractMockingIntegrationTestBase {
+
+    private static String originalTlsProtocols;
+
+    @BeforeClass
+    public static void startServer() {
+        // save original value
+        originalTlsProtocols = tlsProtocols();
+
+        // set new certificate authority values
+        tlsProtocols("TLSv1.2,TLSv1.3");
+
+        mockServerClient = ClientAndServer.startClientAndServer().withSecure(true);
+    }
+
+    @AfterClass
+    public static void stopServer() {
+        stopQuietly(mockServerClient);
+
+        // set back to original value
+        tlsProtocols(originalTlsProtocols);
+    }
+
+    @Override
+    public int getServerPort() {
+        return mockServerClient.getPort();
+    }
+
+    private static void supportsTLSv1_3() throws NoSuchAlgorithmException {
+        assumeThat(
+            "TLSv1.3 is not supported in JDK",
+            SSLContext.getDefault().getSupportedSSLParameters().getProtocols(),
+            hasItemInArray("TLSv1.3")
+        );
+    }
+
+    @Test
+    public void shouldReturnResponseInHTTP() throws NoSuchAlgorithmException {
+        // given
+        supportsTLSv1_3();
+
+        // when
+        mockServerClient
+            .when(
+                request()
+                    .withMethod("POST")
+                    .withPath(calculatePath("some_path"))
+            )
+            .respond(
+                response()
+                    .withStatusCode(200)
+                    .withBody("some_body_response")
+            );
+
+        // then
+        // - in http
+        assertEquals(
+            response()
+                .withStatusCode(OK_200.code())
+                .withReasonPhrase(OK_200.reasonPhrase())
+                .withBody("some_body_response"),
+            makeRequest(
+                request()
+                    .withPath(calculatePath("some_path"))
+                    .withMethod("POST"),
+                getHeadersToRemove()
+            )
+        );
+    }
+
+    @Test
+    public void shouldReturnResponseInHTTPS() throws NoSuchAlgorithmException {
+        // given
+        supportsTLSv1_3();
+
+        // when
+        mockServerClient
+            .when(
+                request()
+                    .withMethod("POST")
+                    .withPath(calculatePath("some_path"))
+            )
+            .respond(
+                response()
+                    .withStatusCode(200)
+                    .withBody("some_body_response")
+            );
+
+        // then
+        // - in https
+        assertEquals(
+            response()
+                .withStatusCode(OK_200.code())
+                .withReasonPhrase(OK_200.reasonPhrase())
+                .withBody("some_body_response"),
+            makeRequest(
+                request()
+                    .withSecure(true)
+                    .withPath(calculatePath("some_path"))
+                    .withMethod("POST"),
+                getHeadersToRemove()
+            )
+        );
+    }
+
+    @Test
+    public void shouldForwardRequestInHTTPS() throws NoSuchAlgorithmException {
+        // given
+        supportsTLSv1_3();
+
+        // when
+        Expectation[] upsertedExpectations = mockServerClient
+            .when(
+                request()
+                    .withPath(calculatePath("echo"))
+            )
+            .forward(
+                forward()
+                    .withHost("127.0.0.1")
+                    .withPort(secureEchoServer.getPort())
+                    .withScheme(HttpForward.Scheme.HTTPS)
+            );
+        // then
+        assertThat(upsertedExpectations.length, is(1));
+        assertThat(upsertedExpectations[0], is(
+            new Expectation(
+                request()
+                    .withPath(calculatePath("echo"))
+            )
+                .thenForward(
+                    forward()
+                        .withHost("127.0.0.1")
+                        .withPort(secureEchoServer.getPort())
+                        .withScheme(HttpForward.Scheme.HTTPS)
+                )
+        ));
+
+        // then
+        // - in http
+        assertEquals(
+            response()
+                .withStatusCode(OK_200.code())
+                .withReasonPhrase(OK_200.reasonPhrase())
+                .withHeaders(
+                    header("x-test", "test_headers_and_body")
+                )
+                .withBody("an_example_body_http"),
+            makeRequest(
+                request()
+                    .withPath(calculatePath("echo"))
+                    .withMethod("POST")
+                    .withHeaders(
+                        header("x-test", "test_headers_and_body")
+                    )
+                    .withBody("an_example_body_http"),
+                getHeadersToRemove()
+            )
+        );
+        // - in https
+        assertEquals(
+            response()
+                .withStatusCode(OK_200.code())
+                .withReasonPhrase(OK_200.reasonPhrase())
+                .withHeaders(
+                    header("x-test", "test_headers_and_body")
+                )
+                .withBody("an_example_body_https"),
+            makeRequest(
+                request()
+                    .withSecure(true)
+                    .withPath(calculatePath("echo"))
+                    .withMethod("POST")
+                    .withHeaders(
+                        header("x-test", "test_headers_and_body")
+                    )
+                    .withBody("an_example_body_https"),
+                getHeadersToRemove()
+            )
+        );
+    }
+
+
+}